New `load_historical` command for pulling detached records from Data Broker DB
Logic is pulled from the `load_submission` command into a `load_base` module,
from which it is used by `load_historical` to get historical contract info
outside the usual by-submission load.
