+++ usaspending_api/etl/broker_etl_helpers.py
@@ -18,9 +18,8 @@ class PhonyCursor:
     """Spoofs the db cursor responses."""
 
     def __init__(self):
-        json_data = open(os.path.join(os.path.dirname(__file__), 'tests/etl_test_data.json'))
-        self.db_responses = json.load(json_data)
-        json_data.close()
+        with open(os.path.join(os.path.dirname(__file__), 'tests/etl_test_data.json')) as json_data:
+            self.db_responses = json.load(json_data)
 
         self.results = None
