+++ usaspending_api/etl/tests/test_load_submission.py
@@ -36,6 +36,26 @@ def partially_flushed():
 
 
 @pytest.mark.django_db
+def test_load_historical_command(endpoint_data, partially_flushed):
+    """
+    Test the submission loader to validate the ETL process
+    """
+    # Load the RefObjClass and ProgramActivityCode data
+    assert Award.objects.count() == 0
+    assert TransactionContract.objects.count() == 0
+    assert LegalEntity.objects.count() == 0
+    assert Location.objects.count() == 0
+    call_command('load_historical', '--test',
+                 '--action_date_begin', '2017-05-01',
+                 '--action_date_end', '2017-05-02',
+                 '--awarding_agency_code', '015')
+    assert Award.objects.count() == 9
+    assert TransactionContract.objects.count() == 10
+    assert LegalEntity.objects.count() == 9
+    assert Location.objects.count() == 10
+
+
+@pytest.mark.django_db
 def test_load_submission_command(endpoint_data, partially_flushed):
     """
     Test the submission loader to validate the ETL process
